Change X509_get_pathlen return to int32_t
Also rename test/v3ext.c to test/v3ext_test.c

+++ crypto/include/internal/x509_int.h
@@ -147,7 +147,7 @@ struct x509_st {
     int references;
     CRYPTO_EX_DATA ex_data;
     /* These contain copies of various extension values */
-    long ex_pathlen;
+    int32_t ex_pathlen;
     long ex_pcpathlen;
     uint32_t ex_flags;
     uint32_t ex_kusage;
